Making app work right through fragment
