--- app/src/main/java/com/javernaut/gquiz/MainFragment.java
@@ -1,6 +1,0 @@
-package com.javernaut.gquiz;
-
-import androidx.fragment.app.Fragment;
-
-public class MainFragment extends Fragment {
-}
